+++ Youhak/src/main/java/net/youhak/member/entity/RoleCategory.java
@@ -1,4 +1,4 @@
-package net.youhak.member.role.domain;
+package net.youhak.member.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.OneToMany;
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import java.util.List;
 import lombok.Builder;
 import lombok.RequiredArgsConstructor;
-import net.youhak.global.BaseEntity;
+import net.youhak.shared.entity.BaseEntity;
 
 @Entity
 @RequiredArgsConstructor
