+++ Youhak/src/main/java/net/youhak/member/repository/RoleJpaRepository.java
@@ -1,6 +1,6 @@
-package net.youhak.member.role.repository;
+package net.youhak.member.repository;
 
-import net.youhak.member.role.domain.Role;
+import net.youhak.member.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
