+++ Youhak/src/main/java/net/youhak/member/repository/RoleRepository.java
@@ -1,7 +1,7 @@
-package net.youhak.member.role.repository;
+package net.youhak.member.repository;
 
-import net.youhak.member.role.domain.Role;
-import net.youhak.member.role.domain.RoleCategory;
+import net.youhak.member.entity.Role;
+import net.youhak.member.entity.RoleCategory;
 import org.springframework.stereotype.Repository;
 
 @Repository
